+++ CHANGELOG.md
@@ -27,11 +27,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added a new ``--bos-templates`` command-line option and a ``bos_templates``
+  configuration file option to ``sat bootsys boot|shutdown --stage bos-operations``.
+  This is intended to replace ``--cle-bos-template``, ``--uan-bos-template``,
+  and their respective configuration file options, ``cle_bos_template`` and
+  ``uan_bos_template``.
+
 ### Changed
+- ``--cle-bos-template`` and ``--uan-bos-template`` now no longer have defaults
+  for ``sat bootsys boot|shutdown --stage bos-operations``. If a CLE or UAN
+  session template is not specified then it will not be used for the boot or
+  shutdown.
 - Changed the default for s3.endpoint in ``sat.toml``
   from ``https://rgw-vip`` to ``https://rgw-vip.nmn``.
 - Changed ``sat firmware`` to display both 'name' and 'targetName' fields from
   FAS, and changed the 'ID' header in the table to read 'name'.
+
+### Fixed
+- Changed "BIS server" in SAT man page to "Kubernetes manager nodes", reworded
+  description to discuss all of SAT's uses, and corrected several small typos
+  in the SAT man page.
 
 ## [3.3.0] - 2021-02-05